Move get_num_workers to utils
