+++ python/lsst/resources/_resourcePath.py
@@ -19,12 +19,10 @@ import copy
 import io
 import locale
 import logging
-import multiprocessing
 import os
 import posixpath
 import re
 import urllib.parse
-from functools import cache
 from pathlib import Path, PurePath, PurePosixPath
 from random import Random
 
@@ -39,7 +37,7 @@ from collections.abc import Iterable, Iterator
 from typing import TYPE_CHECKING, Any, Literal, NamedTuple, overload
 
 from ._resourceHandles._baseResourceHandle import ResourceHandleProtocol
-from .utils import get_tempdir
+from .utils import _get_num_workers, get_tempdir
 
 if TYPE_CHECKING:
     from .utils import TransactionProtocol
@@ -53,49 +51,12 @@ ESCAPES_RE = re.compile(r"%[A-F0-9]{2}")
 # Precomputed escaped hash
 ESCAPED_HASH = urllib.parse.quote("#")
 
-# Maximum number of worker threads for parallelized operations.
-# If greater than 10, be aware that this number has to be consistent
-# with connection pool sizing (for example in urllib3).
-MAX_WORKERS = 10
-
 
 class MTransferResult(NamedTuple):
     """Report on a bulk transfer."""
 
     success: bool
     exception: Exception | None
-
-
-def _get_int_env_var(env_var: str) -> int | None:
-    int_value = None
-    env_value = os.getenv(env_var)
-    if env_value is not None:
-        with contextlib.suppress(TypeError):
-            int_value = int(env_value)
-    return int_value
-
-
-@cache
-def _get_num_workers() -> int:
-    f"""Calculate the number of workers to use.
-
-    Returns
-    -------
-    num : `int`
-        The number of workers to use. Will use the value of the
-        ``LSST_RESOURCES_NUM_WORKERS`` environment variable if set. Will fall
-        back to using the CPU count (plus 2) but capped at {MAX_WORKERS}.
-    """
-    num_workers: int | None = None
-    num_workers = _get_int_env_var("LSST_RESOURCES_NUM_WORKERS")
-    if num_workers is None:
-        # CPU_LIMIT is used on nublado.
-        cpu_limit = _get_int_env_var("CPU_LIMIT") or multiprocessing.cpu_count()
-        if cpu_limit is not None:
-            num_workers = cpu_limit + 2
-
-    # But don't ever return more than the maximum allowed.
-    return min([num_workers, MAX_WORKERS])
 
 
 class ResourcePath:  # numpydoc ignore=PR02
